Adding toaster when delete a post, and adding also a validation on deletePost API, to check if the user who is deleting the post is the same user who created the post

+++ pages/api/posts/deletePost.tsx
@@ -1,17 +1,40 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 
-import { useQuery } from '@tanstack/react-query';
-
 import prisma from '../../../prisma/client';
 
 export default async function handler(
 	req: NextApiRequest,
 	res: NextApiResponse
 ) {
+	// this postId is the postId of the post to be deleted
+	const { postId } = req.body;
+	const { session } = req.body;
+
+	/**
+	 * Check if the user who is trying to delete the post is the same user who created the post and if is the same user who is logged in
+	 */
+	const postToBeDeleted = await prisma.Post.findUnique({
+		where: {
+			idd: postId,
+		},
+	});
+
+	const userWhoCreatedPost = await prisma.user.findUnique({
+		where: {
+			id: postToBeDeleted.userId,
+		},
+	});
+
+	if (userWhoCreatedPost.email !== session?.user?.email) {
+		return res.status(403).json({
+			err: 'You are not authorized to delete this post',
+		});
+	}
+
 	try {
 		const result = await prisma.Post.delete({
 			where: {
-				id: req.body.id,
+				id: postId,
 			},
 			include: {
 				comments: true,
